pyproject.toml
[tool.poetry]
name = "mondaybot"
version = "0.1.0"
description = ""
authors = ["Alex Martens <alex@thinglab.org>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2"

[tool.poetry.scripts]
mondaybot = "mondaybot.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
